requirements.txt
alembic==1.13.1
amqp==5.2.0
anyio==4.3.0
asgiref==3.7.2
Babel==2.14.0
bcrypt==4.1.2
beautiful-date==2.3.0
billiard==4.2.0
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decouple==0.0.7
dep-logic==0.2.0
distlib==0.3.8
Django==4.2.11
filelock==3.13.1
findpython==0.6.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.49.0
gcsa==2.2.0
google-api-core==2.17.1
google-api-python-client==2.126.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
hishel==0.0.25
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
installer==0.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.3.7
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
msgpack==1.0.8
mysqlclient==2.2.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
passlib==1.7.4
pbs-installer==2024.4.1
pdm==2.13.2
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrtlsdr==0.3.0
pyrtlsdrlib==0.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
resolvelib==1.0.1
rich==13.7.1
rsa==4.9
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
speaklater==1.3
SQLAlchemy==2.0.29
sqlparse==0.4.4
tomlkit==0.12.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==4.3.1
unearth==0.15.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.2
WTForms==3.1.2
zstandard==0.22.0
